+++ channel/room_trade.go
@@ -1,0 +1,279 @@
+package channel
+
+import (
+	"log"
+	"math"
+
+	"github.com/Hucaru/Valhalla/common/opcode"
+	"github.com/Hucaru/Valhalla/constant"
+	"github.com/Hucaru/Valhalla/mpacket"
+)
+
+type tradeRoom struct {
+	room
+	mesos     map[int32]int32
+	items     map[int32]map[byte]Item
+	confirmed map[int32]bool
+
+	finalized bool
+}
+
+func newTradeRoom(id int32) roomer {
+	r := room{roomID: id, roomType: constant.MiniRoomTypeTrade}
+	return &tradeRoom{room: r, mesos: make(map[int32]int32), items: make(map[int32]map[byte]Item), confirmed: make(map[int32]bool), finalized: false}
+}
+
+func (r *tradeRoom) addPlayer(plr *Player) bool {
+	if !r.room.addPlayer(plr) {
+		return false
+	}
+
+	plr.Send(packetRoomShowWindow(r.roomType, constant.MiniRoomTypeTrade, byte(constant.RoomMaxPlayers), byte(len(r.players)-1), "", r.players))
+
+	if len(r.players) > 1 {
+		r.sendExcept(packetRoomJoin(r.roomType, byte(len(r.players)-1), r.players[len(r.players)-1]), plr)
+	}
+
+	r.items[plr.ID] = make(map[byte]Item)
+	r.confirmed[plr.ID] = false
+
+	return true
+}
+
+func (r *tradeRoom) removePlayer(plr *Player) {
+	r.closeWithReason(constant.RoomLeaveTradeCancelled, true)
+}
+
+func (r tradeRoom) sendInvite(plr *Player) {
+	plr.Send(packetRoomInvite(constant.MiniRoomTypeTrade, r.players[0].Name, r.roomID))
+}
+
+func (r tradeRoom) reject(code byte, name string) {
+	r.send(packetRoomInviteResult(code, name))
+}
+
+func (r *tradeRoom) canInsertItem(plrID int32, tradeSlot byte) bool {
+	if tradeSlot < 1 || tradeSlot > 9 {
+		return false
+	}
+	if r.items == nil {
+		return false
+	}
+	if _, ok := r.items[plrID]; !ok || r.items[plrID] == nil {
+		return false
+	}
+	if _, exists := r.items[plrID][tradeSlot]; exists {
+		return false
+	}
+	return true
+}
+
+func (r *tradeRoom) insertItem(tradeSlot byte, plrID int32, item Item) bool {
+	if tradeSlot < 1 || tradeSlot > 9 {
+		log.Printf("trade: invalid slot %d from player %d\n", tradeSlot, plrID)
+		return false
+	}
+
+	if r.items == nil || r.items[plrID] == nil {
+		log.Printf("trade: missing items map for player %d\n", plrID)
+		return false
+	}
+
+	if _, exists := r.items[plrID][tradeSlot]; exists {
+		log.Printf("trade: slot %d already occupied for player %d\n", tradeSlot, plrID)
+		return false
+	}
+
+	r.items[plrID][tradeSlot] = item
+	isUser0 := r.players[0].ID == plrID
+	r.players[0].Send(packetRoomTradePutItem(tradeSlot, !isUser0, item))
+	r.players[1].Send(packetRoomTradePutItem(tradeSlot, isUser0, item))
+	return true
+}
+
+func (r *tradeRoom) updateMesos(amount, plrID int32) {
+	r.mesos[plrID] += amount
+	isUser0 := r.players[0].ID == plrID
+	r.players[0].Send(packetRoomTradePutMesos(r.mesos[plrID], !isUser0))
+	r.players[1].Send(packetRoomTradePutMesos(r.mesos[plrID], isUser0))
+}
+
+func (r *tradeRoom) acceptTrade(plr *Player) bool {
+	r.confirmed[plr.ID] = true
+
+	for _, user := range r.players {
+		if user.ID != plr.ID {
+			user.Send(packetRoomTradeAccept())
+		}
+	}
+
+	if r.confirmed[r.players[0].ID] && r.confirmed[r.players[1].ID] {
+		r.completeTrade()
+	}
+
+	return r.finalized
+}
+
+func (r *tradeRoom) completeTrade() {
+	if len(r.players) < 2 || r.players[0] == nil || r.players[1] == nil {
+		r.closeWithReason(constant.MiniRoomTradeFail, true)
+		return
+	}
+
+	p1 := r.players[0]
+	p2 := r.players[1]
+
+	type tradeChange struct {
+		plr         *Player
+		mesosChange int32
+		itemsToGive []Item
+	}
+
+	changes := []tradeChange{
+		{plr: p1, mesosChange: r.mesos[p2.ID]},
+		{plr: p2, mesosChange: r.mesos[p1.ID]},
+	}
+
+	for _, item := range r.items[p1.ID] {
+		changes[1].itemsToGive = append(changes[1].itemsToGive, item)
+	}
+
+	for _, item := range r.items[p2.ID] {
+		changes[0].itemsToGive = append(changes[0].itemsToGive, item)
+	}
+
+	if int64(p1.mesos)+int64(changes[0].mesosChange) > int64(math.MaxInt32) ||
+		int64(p2.mesos)+int64(changes[1].mesosChange) > int64(math.MaxInt32) ||
+		changes[0].mesosChange < 0 || changes[1].mesosChange < 0 {
+		r.closeWithReason(constant.MiniRoomTradeFail, true)
+		return
+	}
+
+	if !p1.canReceiveItems(changes[0].itemsToGive) || !p2.canReceiveItems(changes[1].itemsToGive) {
+		r.closeWithReason(constant.MiniRoomTradeInventoryFull, true)
+		return
+	}
+
+	var undo []func()
+	defer func() {
+		if r.finalized {
+			return
+		}
+		for i := len(undo) - 1; i >= 0; i-- {
+			func(fn func()) {
+				defer func() { _ = recover() }()
+				fn()
+			}(undo[i])
+		}
+	}()
+
+	for _, it := range changes[0].itemsToGive {
+		err, gi := p1.GiveItem(it)
+		if err != nil {
+			log.Printf("Trade error: failed to give item %v to %s: %v", it.ID, p1.Name, err)
+			r.closeWithReason(constant.MiniRoomTradeInventoryFull, true)
+			return
+		}
+
+		undo = append(undo, func() {
+			if _, err := p1.takeItem(gi.ID, gi.slotID, gi.amount, gi.invID); err != nil {
+				log.Printf("Trade rollback warning: failed to remove item %v from %s: %v", gi.ID, p1.Name, err)
+			}
+		})
+	}
+
+	for _, it := range changes[1].itemsToGive {
+		err, gi := p2.GiveItem(it)
+		if err != nil {
+			log.Printf("Trade error: failed to give item %v to %s: %v", it.ID, p2.Name, err)
+			r.closeWithReason(constant.MiniRoomTradeInventoryFull, true)
+			return
+		}
+
+		undo = append(undo, func() {
+			if _, err := p2.takeItem(gi.ID, gi.slotID, gi.amount, gi.invID); err != nil {
+				log.Printf("Trade rollback warning: failed to remove item %v from %s: %v", gi.ID, p2.Name, err)
+			}
+		})
+	}
+
+	if changes[0].mesosChange > 0 {
+		mc := changes[0].mesosChange
+		p1.giveMesos(mc)
+		undo = append(undo, func() { p1.giveMesos(-mc) })
+	}
+
+	if changes[1].mesosChange > 0 {
+		mc := changes[1].mesosChange
+		p2.giveMesos(mc)
+		undo = append(undo, func() { p2.giveMesos(-mc) })
+	}
+
+	r.finalized = true
+	r.closeWithReason(constant.MiniRoomTradeSuccess, false)
+}
+
+func (r *tradeRoom) closeWithReason(reason byte, rollback bool) {
+	if rollback {
+		r.rollback()
+	}
+	for i, plr := range r.players {
+		if plr != nil {
+			plr.Send(packetRoomLeave(byte(i), reason))
+		}
+	}
+}
+
+func (r *tradeRoom) rollback() {
+	if r.finalized {
+		return
+	}
+
+	for _, player := range r.players {
+		if player == nil {
+			continue
+		}
+		if m := r.mesos[player.ID]; m != 0 {
+			player.giveMesos(m)
+			r.mesos[player.ID] = 0
+		}
+		if bag, ok := r.items[player.ID]; ok {
+			for slot, item := range bag {
+				if err, _ := player.GiveItem(item); err != nil {
+					log.Println("tradeRoom rollback failed:", err)
+				}
+				delete(bag, slot)
+			}
+		}
+	}
+
+	r.finalized = true
+}
+
+func packetRoomTradePutItem(tradeSlot byte, user bool, item Item) mpacket.Packet {
+	p := mpacket.CreateWithOpcode(opcode.SendChannelRoom)
+	p.WriteByte(constant.MiniRoomTradePutItem)
+	p.WriteBool(user)
+	p.WriteByte(tradeSlot)
+	p.WriteBytes(item.StorageBytes())
+	return p
+}
+
+func packetRoomTradePutMesos(amount int32, user bool) mpacket.Packet {
+	p := mpacket.CreateWithOpcode(opcode.SendChannelRoom)
+	p.WriteByte(constant.MiniRoomTradePutMesos)
+	p.WriteBool(user)
+	p.WriteInt32(amount)
+	return p
+}
+
+func packetRoomTradeAccept() mpacket.Packet {
+	p := mpacket.CreateWithOpcode(opcode.SendChannelRoom)
+	p.WriteByte(constant.MiniRoomTradeAccept)
+
+	return p
+}
+
+func packetRoomTradeRequireSameMap() mpacket.Packet {
+	return packetRoomEnterErrorMsg(constant.RoomEnterTradeSameMap)
+}
